CRUD complet Piece

+++ FileRouge/src/main/java/groupe1/filrouge/entity/Vehicule.java
@@ -29,7 +29,7 @@ public class Vehicule {
 	private Integer qte;
 
 	@Column(name = "prixHT")
-	private double prixHT;
+	private Float prixHT;
 
 	@Temporal(TemporalType.DATE)
 	@Column(name = "date_creation", nullable = false)
@@ -70,11 +70,11 @@ public class Vehicule {
 		this.qte = quantite;
 	}
 
-	public double getPrixHT() {
+	public Float getPrixHT() {
 		return prixHT;
 	}
 
-	public void setPrixHT(double prixHT) {
+	public void setPrixHT(Float prixHT) {
 		this.prixHT = prixHT;
 	}
